deepfake_api/requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-image>=0.20.0
face-recognition>=1.3.0
mtcnn>=0.1.1
facenet-pytorch>=2.5.2
insightface>=0.7.3

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
scipy>=1.10.0
numpy>=1.24.0

# Video Processing
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
httpx>=0.24.0

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
aiofiles>=23.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0

# Model specific
timm>=0.9.0
efficientnet-pytorch>=0.7.1
huggingface-hub>=0.16.0

# Additional ML utilities
scikit-learn>=1.3.0
joblib>=1.3.0